# src/ideapress/web/routes/projects.py
"""ideapress.web.routes.projects — the project API and the project pages.

Route handlers contain no business logic: each calls one service method and renders. Model-produced
and user-supplied text reaches the templates as plain strings and is escaped once by the shared
macros — never ``| safe`` (risk S1).
"""

from __future__ import annotations

import base64
import json
from typing import TYPE_CHECKING, Annotated, Any

from baseaicore import ValidationError
from fastapi import APIRouter, Form, Query, Request, status
from mirrorwall import json_response, paginated_response
from pydantic import BaseModel, ConfigDict, Field

# Imported at runtime, not under TYPE_CHECKING: FastAPI reads a handler's return annotation
# when it builds the OpenAPI schema, and a forward reference it cannot resolve makes
# `app.openapi()` raise — which is a 500 on /api/v1/docs that no other test would notice.
from starlette.responses import HTMLResponse, JSONResponse, RedirectResponse, Response

from ideapress.web.csrf import render_form_page

if TYPE_CHECKING:
    from ideapress.domain.project import Project
    from ideapress.services.projects import ProjectService

__all__ = ["router", "ui_router"]

router = APIRouter(tags=["projects"])
ui_router = APIRouter(include_in_schema=False)


class CreateProjectRequest(BaseModel):
    """``POST /projects`` body."""

    model_config = ConfigDict(extra="forbid")

    title: str = Field(min_length=1, max_length=200)
    brief: str = ""
    content_type: str = "article"
    workflow_id: str = "standard"
    author_material: dict[str, Any] = Field(default_factory=dict)


class UpdateProjectRequest(BaseModel):
    """``PUT /projects/{id}`` body. Every field optional; omitted fields are untouched."""

    model_config = ConfigDict(extra="forbid")

    title: str | None = Field(default=None, max_length=200)
    brief: str | None = None
    author_material: dict[str, Any] | None = None
    status: str | None = None


def _service(request: Request) -> ProjectService:
    """Return the project service this process built, from the runtime the lifespan opened."""
    return request.app.state.runtime.projects  # type: ignore[no-any-return]  # set by build_runtime


def _as_payload(project: Project) -> dict[str, Any]:
    """Render a project as the API's JSON shape."""
    return {
        "id": project.id,
        "title": project.title,
        "slug": project.slug,
        "content_type": project.content_type,
        "content_type_version": project.content_type_version,
        "workflow_id": project.workflow_id,
        "workflow_version": project.workflow_version,
        "status": project.status,
        "brief": project.brief_text,
        "author_material": project.author_material,
        "created_at": project.created_at.isoformat(),
        "updated_at": project.updated_at.isoformat(),
        "completed_at": project.completed_at.isoformat() if project.completed_at else None,
        "archived_at": project.archived_at.isoformat() if project.archived_at else None,
    }


@router.post("/projects", status_code=status.HTTP_201_CREATED)
def create_project(request: Request, body: CreateProjectRequest) -> JSONResponse:
    """Create a project. The slug is derived from the title, never supplied by the caller."""
    project = _service(request).create(
        title=body.title,
        brief=body.brief,
        content_type=body.content_type,
        workflow_id=body.workflow_id,
        author_material=body.author_material,
    )
    return json_response(_as_payload(project), status=status.HTTP_201_CREATED)


@router.get("/projects")
def list_projects(
    request: Request,
    project_status: Annotated[str | None, Query(alias="status")] = None,
    content_type: Annotated[str | None, Query()] = None,
    include_archived: Annotated[bool, Query()] = False,
    limit: Annotated[int, Query(ge=1, le=200)] = 50,
    offset: Annotated[int, Query(ge=0)] = 0,
    cursor: Annotated[str | None, Query()] = None,
) -> JSONResponse:
    """List projects, most recently updated first. Archived projects are hidden by default.

    ``cursor`` is the previous page's ``next_cursor`` (API standards §6) and wins over ``offset``,
    which stays for callers that already send it.

    Raises:
        ValidationError: ``cursor`` is not one this endpoint issued; ``400`` naming ``cursor``.
    """
    start = _offset_of(cursor) if cursor else offset
    projects = _service(request).list(
        status=project_status,
        content_type=content_type,
        include_archived=include_archived,
        limit=limit + 1,
        offset=start,
    )
    has_more = len(projects) > limit
    page = [_as_payload(project) for project in projects[:limit]]
    return paginated_response(
        page,
        limit=limit,
        has_more=has_more,
        next_cursor=_cursor_at(start + limit) if has_more else None,
    )


def _cursor_at(offset: int) -> str:
    """The opaque cursor for the page that starts at ``offset``: API standards §6's own shape."""
    # ponytail: an offset cursor, so a project updated between two page reads can cross a page
    # edge; key it on (updated_at, id) if the list is ever paged under concurrent edits.
    return base64.urlsafe_b64encode(json.dumps({"offset": offset}).encode()).decode()


def _offset_of(cursor: str) -> int:
    """The offset a cursor from :func:`_cursor_at` names.

    Raises:
        ValidationError: The cursor is not one this endpoint issued — refused by name rather than
            read as the first page, which would hand a paging loop that page forever.
    """
    try:
        offset = json.loads(base64.urlsafe_b64decode(cursor.encode()))["offset"]
    except (ValueError, TypeError, KeyError):
        offset = None
    if not isinstance(offset, int) or isinstance(offset, bool) or offset < 0:
        message = "cursor is not one this endpoint issued; start again without it."
        raise ValidationError(
            message, details={"fields": [{"path": "cursor", "problem": "not an issued cursor"}]}
        )
    return offset


@router.get("/projects/{project_id}")
def get_project(request: Request, project_id: str) -> JSONResponse:
    """Return one project, its plan summary, its unit states and its stage history (api.md §2).

    Raises:
        ProjectNotFound: Rendered as 404 by the shared handler.
    """
    from ideapress.services.stage_reports import project_overview

    project = _service(request).get(project_id)
    overview = project_overview(request.app.state.runtime, project_id=project_id)
    return json_response({**_as_payload(project), **overview})


@router.get("/projects/{project_id}/research")
def get_research(request: Request, project_id: str) -> JSONResponse:
    """Where a fetch may go, and every note and tool call the research stage left (row WP5).

    The stage starts at ``POST /projects/{id}/stages/research/run``. This is what it may reach
    before anyone starts it (spec §14: a click that may fetch says where it can go) and what it
    recorded afterwards, each call with the egress decision it ran under.

    Raises:
        ProjectNotFound: No such project.
    """
    from ideapress.services.unit_reports import research_report

    _service(request).get(project_id)
    runtime = request.app.state.runtime
    research = runtime.settings.research
    return json_response(
        {
            "allowed_hosts": list(research.allowed_hosts),
            "allowed_tools": list(research.allowed_tools),
            **research_report(runtime, project_id=project_id),
        }
    )


@router.put("/projects/{project_id}")
def update_project(request: Request, project_id: str, body: UpdateProjectRequest) -> JSONResponse:
    """Update the brief, author material, title or status. Never recompiles requirements."""
    project = _service(request).update(
        project_id,
        title=body.title,
        brief=body.brief,
        author_material=body.author_material,
        status=body.status,
    )
    return json_response(_as_payload(project))


@router.delete("/projects/{project_id}")
def delete_project(
    request: Request,
    project_id: str,
    confirm: Annotated[bool, Query()] = False,
    archive: Annotated[bool, Query()] = False,
) -> JSONResponse:
    """Preview or perform a delete, archiving the project first when asked.

    Without ``?confirm=true`` this removes nothing and returns what *would* be removed. That is
    api.md §2's preview-then-confirm, and it is the difference between a mis-click and the loss of
    the user's own writing. ``?archive=true`` with it writes the project's archive first, into
    ``archive_directory``; an archive that cannot be written deletes nothing (row WP5).

    Raises:
        ProjectNotFound: No such project.
        ExportFailed: The archive could not be written; nothing was deleted.
    """
    from ideapress.services.project_archive import delete_project as delete_with_archive

    result = delete_with_archive(
        request.app.state.runtime, project_id=project_id, confirm=confirm, archive=archive
    )
    preview = result["preview"]
    return json_response(
        {
            "deleted": result["deleted"],
            "project": _as_payload(preview.project),
            "source_count": preview.source_count,
            "directory": str(preview.directory) if preview.directory else None,
            "directory_bytes": preview.directory_bytes,
            "archive_directory": result["archive_directory"],
            "archive": result["archive"],
        }
    )


@ui_router.get("/")
def projects_page(
    request: Request, show_archived: Annotated[bool, Query()] = False
) -> HTMLResponse:
    """Render the project list."""
    projects = _service(request).list(include_archived=show_archived)
    return render_form_page(
        request,
        "projects/index.html",
        page="projects",
        page_title="Projects",
        projects=projects,
        show_archived=show_archived,
    )


@ui_router.get("/projects/{project_id}")
def project_page(request: Request, project_id: str) -> HTMLResponse:
    """Render one project's detail shell."""
    project = _service(request).get(project_id)
    return render_form_page(
        request,
        "projects/detail.html",
        page="projects",
        page_title=project.title,
        project=project,
    )


@ui_router.post("/projects")
def create_project_form(
    request: Request,
    title: Annotated[str, Form()],
    brief: Annotated[str, Form()] = "",
    content_type: Annotated[str, Form()] = "article",
) -> Response:
    """Create a project from the list page's form, then redirect to it."""
    project = _service(request).create(title=title, brief=brief, content_type=content_type)
    return RedirectResponse(f"/projects/{project.id}", status_code=status.HTTP_303_SEE_OTHER)